refactor common testing code to a common place
